txn/tests: check that txes are fully applied
Redpanda transactions have read committed isolation level so
they may ack a commit request before it's fully committed. In
this case the next operation will finish the cleanup.

If the cleanup fails it fails the next test while in fact it's
a delayed error of the first test. Adding an empty transaction
after each test to trigger the clean up phase and be sure that
a tx is fully executed before jumping to the text test.